Iterration 1 of Rewriting Game Engine to allow multithreading updates
