+++ src/Main.java
@@ -1,5 +1,7 @@
+import Engine_v2.Engine;
+
 public class Main {
     public static void main(String[] args){
-        new Engine.GUI();
+        new Engine();
     }
 }
